[REMOVE] Rank 관련 작성 코드 제거 및 이동
